Add package-lock.json file
The package-lock.json file was added. This file holds information about our project dependencies and ensures consistent installations and builds across environments. It contains detailed entries for each package this project depends on, including the specific versions that should be used. This not only helps handle dependencies more effectively but also improves the security by locking the versions of the sub-dependencies being used.