Implement significant frontend components and routing
This commit introduces several frontend components such as Navbar, Home, Footer, and many other Vue components. It also adds routing in Vue to handle page navigation, creates vuex store for handling the state of the application, and adjusts the configurations for Laravel Vite Plugin, PostCSS, and TailwindCSS. Other changes include modification in the .env file and addition of the .env.local file for improved security setup.