Update package versions in composer.lock file
The composer.lock file has been updated to reflect the latest versions of certain packages including laravel/framework, nesbot/carbon, phpunit/phpunit, and spatie/ignition. This change will help to ensure that the project is using the most up-to-date and stable versions of these packages for optimal performance and security.