Implement Spatie/laravel-permission for role management
The commit replaces the previous custom role system with the Spatie/laravel-permission package, a more flexible and efficient solution. This change entailed deleting the roles and role_user table migrations and the RoleSeeder, and adding the Spatie package to composer files and User model. New migration and seeder were also added to create necessary tables for working with roles and permissions using the Spatie package.